Implemented grave digging
