+++ info.json
@@ -3,5 +3,5 @@
     "name": "Eat The Dead",
     "description": "No pork? No problem! Dead villagers now provide meat, which can be your temporary source early in the game, or permanently if you're a damn savage.",
     "author": "cmjten10",
-    "version": "1" 
+    "version": "1.1" 
 }
